DAO-176 Handle loading state, move code out from claim-actions, and cleanup
