+++ src/pages/claim-details/claim-actions.module.scss
@@ -48,60 +48,3 @@
   font-weight: 400;
   max-width: 50ch;
 }
-
-.cost {
-  font-size: $text-xlarge;
-  line-height: $lh-heading-4;
-  font-weight: 600;
-  margin-bottom: 3rem;
-}
-
-.confirmationBody {
-  text-align: center;
-}
-
-.info {
-  margin-bottom: 1rem;
-}
-
-@keyframes loading-skeleton {
-  100% {
-    transform: translateX(100%);
-  }
-}
-
-.skeleton {
-  --base-color: #141514;
-  --highlight-color: #3d3d3b;
-  --animation-duration: 1.5s;
-  --animation-direction: normal;
-
-  background-color: var(--base-color);
-
-  border-radius: 0.25rem;
-  display: inline-flex;
-  line-height: 0.85;
-
-  position: relative;
-  overflow: hidden;
-  z-index: 1; /* Necessary for overflow: hidden to work correctly in Safari */
-}
-
-.skeleton::after {
-  content: ' ';
-  display: block;
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  height: 100%;
-  background-repeat: no-repeat;
-  background-image: linear-gradient(90deg, var(--base-color), var(--highlight-color), var(--base-color));
-  transform: translateX(-100%);
-
-  animation-name: loading-skeleton;
-  animation-direction: var(--animation-direction);
-  animation-duration: var(--animation-duration);
-  animation-timing-function: ease-in-out;
-  animation-iteration-count: infinite;
-}
