+++ src/utils/messages.ts
@@ -40,6 +40,8 @@ export const messages = {
   FAILED_TO_LOAD_CLAIMS: 'Failed to load claims. Please try again later.',
   FAILED_TO_LOAD_POLICIES: 'Failed to load policies. Please try again later.',
   FAILED_TO_LOAD_REMAINING_COVERAGE: 'Failed to load remaining policy coverage. Please try again later.',
+  FAILED_TO_LOAD_ESCALATION_COST: 'Failed to load escalation cost. Please try again later.',
+  FAILED_TO_LOAD_APPEAL_COST: 'Failed to load appeal cost. Please try again later.',
 };
 
 // TODO: these messages should change depending on the final designs
